refactor: Split test files and clean up formatting
- Split net-symbol-refactor.test.tsx into three separate test files
- Each test file now contains only one test function as required
- Clean up formatting and remove unnecessary characters
- All tests pass and functionality remains intact
